Use Snap instead of WAI
The former makes it easier to combine multiple applications (snap handlers)
and so implement fallback chains. We need it if we want to allow users to
serve their own files from a directory, or define more complex routing
setups.

But the more interesting change is that the index HTML is generated by
Haskell, and we can thus allow users to inject arbitrary HTML
into <head>. This is needed to support global CSS (eg. border-box)
and also if users want to load third-party JavaScript libraries
and then use them through FFI.