Replace playground with varna
Varna consists of some small pieces extracted from one of my internal projects.
I've used this for a while to see where the general Nauva API is insufficient
and needs improvement. In particular, the work on nauva-css was largely driven
by the needs of Varna.

Also, this is the first app which features the catalog. Run 'catalogMain' instead
of 'main' (or ':main') to start the catalog. Native builds are not supported yet.